reorganized client modules
