+++ voxels/src/client/gfx/chunk/mod.rs
@@ -1,0 +1,12 @@
+mod s_init;
+mod r_gfx;
+mod vert;
+mod mesh;
+
+pub use self::
+{
+    s_init::*,
+    r_gfx::*,
+    vert::*,
+    mesh::*,
+};
