+++ voxels/src/client/gfx/mod.rs
@@ -1,0 +1,10 @@
+mod s_render;
+mod shared;
+mod chunk;
+
+pub use self::
+{
+    s_render::*,
+    shared::*,
+    chunk::*,
+};
