+++ voxels/src/client/gfx/shared/mod.rs
@@ -1,0 +1,8 @@
+mod s_init;
+mod r_gfx;
+
+pub use self::
+{
+    s_init::*,
+    r_gfx::*,
+};
