--- voxels/src/client/mesh/mod.rs
@@ -1,3 +1,0 @@
-mod sys;
-
-pub use sys::*;
