--- voxels/src/client/render/mod.rs
@@ -1,5 +1,0 @@
-mod res;
-mod sys;
-
-pub use res::*;
-pub use sys::*;
