--- voxels/src/client/render/res/mod.rs
@@ -1,5 +1,0 @@
-mod shared;
-mod chunk;
-
-pub use shared::*;
-pub use chunk::*;
