--- voxels/src/client/render/sys/mod.rs
@@ -1,5 +1,0 @@
-mod render;
-mod init;
-
-pub use render::*;
-pub use init::*;
